feat(ui): updates
