+++ src/app/layout.tsx
@@ -1,7 +1,6 @@
 import './styles/globals.css'
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
-import Navbar from './components/navbar'
 import Footer from './components/footer'
 
 const inter = Inter({ subsets: ['latin'] })
@@ -19,7 +18,6 @@ export default function RootLayout({
   return (
     <html className="scroll-smooth" lang="en">
       <body className={inter.className}>
-        <Navbar/>
         {children}
         <Footer/>
         </body>
